+++ src/config/document.ts
@@ -1,0 +1,6 @@
+module.exports = {
+	rules: {
+		'missing-doctype': 'error',
+		'input-missing-label': 'error',
+	},
+};